+++ Changelog.md
@@ -1,3 +1,8 @@
+## 02.10.2024 - Version 2.8.0
+
+- Add new filters options for `GetTopAnimeAsync` and `GetTopMangaAsync`
+- Fix vulnerability dependancy with System.Text.Json in netstandard2.0 package 
+
 ## 17.11.2023 - Version 2.7.0
 
 - Add support for .net 8